Add Content-Length header for pre-compressed files
We only set it when we know the whole file will be sent
because no Range header was on the request. Otherwise we
defer to the Range mechanics of the Go standard library.
Also set the Vary header when setting Content-Encoding.

The rules are that Content-Length can be set with
Content-Encoding but not with Transfer-Encoding.

The Go Standard library does not set Content-Length when
Content-Encoding is set as a precautionary measure.